better naming and moved some functions to converterService
