+++ src/test/scala/no/ndla/articleapi/service/WriteServiceTest.scala
@@ -105,59 +105,4 @@ class WriteServiceTest extends UnitSuite with TestEnvironment {
     verify(articleIndexService, times(1)).deleteDocument(any[Long])
     verify(searchApiClient, times(1)).deleteArticle(any[Long])
   }
-
-  test("That partialArticleUpdate updates all fields") {
-    val existingArticle = TestData.sampleDomainArticle.copy(
-      grepCodes = Seq("old", "code"),
-      copyright = Copyright("CC-BY-4.0", "origin", Seq(), Seq(), Seq(), None, None, None),
-      metaDescription = Seq(ArticleMetaDescription("gammelDesc", "nb")),
-      tags = Seq(ArticleTag(Seq("gammel", "Tag"), "nb"))
-    )
-    val partialArticle =
-      api.PartialPublishArticle(
-        grepCodes = Some(Seq("New", "grep", "codes")),
-        license = Some("newLicense"),
-        metaDescription = Some(Seq(ArticleMetaDescription("nyDesc", "nb"))),
-        tags = Some(Seq(ArticleTag(Seq("nye", "Tags"), "nb")))
-      )
-    val updatedArticle = TestData.sampleDomainArticle.copy(
-      grepCodes = Seq("New", "grep", "codes"),
-      copyright = Copyright("newLicense", "origin", Seq(), Seq(), Seq(), None, None, None),
-      metaDescription = Seq(ArticleMetaDescription("nyDesc", "nb")),
-      tags = Seq(ArticleTag(Seq("nye", "Tags"), "nb"))
-    )
-
-    service.partialArticleUpdate(existingArticle, partialArticle) should be(updatedArticle)
-
-  }
-
-  test("That partialArticleUpdate does not create new fields") {
-    val existingArticle = TestData.sampleDomainArticle.copy(
-      grepCodes = Seq("old", "code"),
-      copyright = Copyright("CC-BY-4.0", "origin", Seq(), Seq(), Seq(), None, None, None),
-      metaDescription = Seq(ArticleMetaDescription("oldDesc", "de")),
-      tags = Seq(ArticleTag(Seq("Gluten", "Tag"), "de"))
-    )
-    val partialArticle =
-      api.PartialPublishArticle(
-        grepCodes = Some(Seq("New", "grep", "codes")),
-        license = Some("newLicense"),
-        metaDescription = Some(
-          Seq(ArticleMetaDescription("nyDesc", "nb"),
-              ArticleMetaDescription("newDesc", "en"),
-              ArticleMetaDescription("neuDesc", "de"))),
-        tags = Some(
-          Seq(ArticleTag(Seq("nye", "Tags"), "nb"),
-              ArticleTag(Seq("new", "Tagss"), "en"),
-              ArticleTag(Seq("Guten", "Tag"), "de")))
-      )
-    val updatedArticle = TestData.sampleDomainArticle.copy(
-      grepCodes = Seq("New", "grep", "codes"),
-      copyright = Copyright("newLicense", "origin", Seq(), Seq(), Seq(), None, None, None),
-      metaDescription = Seq(ArticleMetaDescription("neuDesc", "de")),
-      tags = Seq(ArticleTag(Seq("Guten", "Tag"), "de"))
-    )
-
-    service.partialArticleUpdate(existingArticle, partialArticle) should be(updatedArticle)
-  }
 }
